Changed the structure of the calculator interface, allowing for selection in the future.
